requirements.txt
discord~=1.7.3
loguru~=0.5.3
requests
python-dotenv~=0.19.2